fix #32: ensure caches are persisted
Previously, the caches were re-initialized on every plugin
initialization. This caused an incompatibility with other tools like
rebundler or persistify.

This change keeps the original intent of creating caches when the plugin
is initialized, but ensures the process keeps the caches instead of the
plugin instance itself. It’s not really possible to have a cache
conflict since they’re keyed by file.