fix a race-condition with tokensByFile
Because we’re setting `tokensByFile` before and after an async action,
and many of these actions can happen at the same time, there’s a race
condition where the incorrect `tokensByFile` can be set.

This accomplishes the same desire for caching by just keeping the loader
for every file since the loader already caches `tokensByFile`, and
prevents the race condition.